+++ VKApiTools/VKApiMessagesDump/VKApiToolsDialog.cpp
@@ -143,7 +143,7 @@ void PrintMessage(const string& Msg, URGB Color)
 	if (Msg.length() <= 0)
 		return;
 	WaitForSingleObject(hMessageMutex, INFINITE);
-	TextWrite LineData(Color, (unsigned short)Msg.length());
+	TextWrite LineData(Color, (uint32_t)Msg.length());
 	dlgS.Edk.AddLines(base91::encode(tw_tostring(LineData)) + " " + Msg);
 	ReleaseMutex(hMessageMutex);
 }
@@ -168,7 +168,7 @@ void PrintDownload(const string& Msg, URGB Color)
 	if (Msg.length() <= 0)
 		return;
 	WaitForSingleObject(hDownloadMutex, INFINITE);
-	TextWrite LineData(Color, (unsigned short)Msg.length());
+	TextWrite LineData(Color, (uint32_t)Msg.length());
 	dlgS.EDownload.AddLines(base91::encode(tw_tostring(LineData)) + " " + Msg);
 	ReleaseMutex(hDownloadMutex);
 }
